// src/shared/api/model/adminUserResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * 프로필 카드 공유 서비스 API
 * 프로필 카드 공유 서비스 백엔드
 * OpenAPI spec version: 0.4.1
 */
import type { FormattedDate } from "./formattedDate";

export interface AdminUserResponse {
  /** 유저 ID */
  id: string;
  /** 유저 이름 */
  name: string;
  /** 유저 닉네임 */
  nickname: string;
  email: string;
  /** 권한 (0: 일반 유저, 1: ADMIN) */
  role: number;
  createdAt: FormattedDate;
  /** 마지막 로그인 시각 (가장 최근 리프레시 토큰 생성 시각) */
  lastLoginAt: FormattedDate | null;
  /** 탈퇴 시각 (미탈퇴 시 null) */
  deletedAt: FormattedDate | null;
}