// src/shared/api/model/formattedDate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * 프로필 카드 공유 서비스 API
 * 프로필 카드 공유 서비스 백엔드
 * OpenAPI spec version: 0.4.1
 */

export interface FormattedDate {
  /** 타임스탬프 (ms) */
  timestamp: number;
  isoString: string;
  /** 상대 시간 */
  timeAgo: string;
}