// src/shared/api/model/usersControllerGetAllUsersParams.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * 프로필 카드 공유 서비스 API
 * 프로필 카드 공유 서비스 백엔드
 * OpenAPI spec version: 0.4.1
 */
import type { UsersControllerGetAllUsersOrder } from "./usersControllerGetAllUsersOrder";

export type UsersControllerGetAllUsersParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * 정렬할 필드명
   */
  sort?: string;
  order?: UsersControllerGetAllUsersOrder;
};